edits to plotting and logging of contrast for baseCon code- now works- LG

+++ BehaviorCode/NewSkeletonCode/MATLAB_window_files/flashingStim_2P_Frames.m
@@ -152,8 +152,8 @@ stimOnFrames = zeros(1, input.maxCyclesOn);
 
 %track contrast by presentation
 multVals = mwGetEventValue(eventsTrial, ds.event_codec, 'multVal', 'all', 1);
-input.tBaseGratingContrast{trN} = input.baseGratingContrast + (multVals .* input.conDiff);
-	 
+input.tBaseGratingContrast{trN} = input.baseGratingContrast + (double(multVals) .* input.conDiff);
+
 auditoryStimOnFrames = zeros(1, input.maxCyclesOn);
 for istim = 1:numberCyclesOn
     try
